Feature/purchase

+++ telegram/src/main/kotlin/ru/justd/cryptobot/IdConverter.kt
@@ -3,4 +3,4 @@ package ru.justd.cryptobot
 
 fun toChannelId(chatId: Long) = chatId.toString()
 
-fun toChatId(channelId: String) = channelId.toLong()
+fun toChatId(channelId: String) = channelId.toInt()
